Fixed linker, need to add opt. for >512b pgms

+++ src/filesystem/initialize_fat.c
@@ -1,4 +1,7 @@
 #include "filesystem/filesystem.h"
+#include "libc/mem.h"
+#include "drivers/ata.h"
+#include "drivers/screen.h"
 
 
 static void initialize_empty_fat_to_disk();
@@ -43,7 +46,7 @@ static void initialize_empty_fat_to_disk() {
    int i = 0;
    for(i = 1; i < rescued_programs_lba[0]; i++) {
       void* program = malloc(512);
-      read_sectors_ATA_PIO(program, rescued_programs_lba[i], 1);
+      read_sectors_ATA_PIO((uint32_t)program, rescued_programs_lba[i], 1);
       struct file* node = fat_head+i;
       memory_copy((uint8_t*)&(((struct program_identifier*)program)->name), (uint8_t*)&(node->name), 32);
       node->lba = rescued_programs_lba[i];
@@ -63,7 +66,7 @@ uint16_t* rescue_program_headers() {
     int i = 0;
     for(i = 0; i < 256; i++) {
         void* program = malloc(512);
-        read_sectors_ATA_PIO(program, i*8, 1);
+        read_sectors_ATA_PIO((uint32_t)program, i*8, 1);
         if(((struct program_identifier*) program)->magic[0] == 0xFFFFFFFF &&
            ((struct program_identifier*) program)->magic[1] == 0xFFFFFFFF &&
            ((struct program_identifier*) program)->magic[2] == 0xFFFFFFFF &&
